Consolidate data stored per-server in LoadBalancer
Locks are also now held more granularly, in that a lock is held for the view
data (that the RendezvousLoadHash will request) as well as for the server
stats, which should only be modified in add_server/load_server (this should be
infrequent) and the background thread itself. The main reason for using a more
granular threading is to provide better performance.